+++ MatrixSDK/Data/MXRoomState.h
@@ -25,6 +25,8 @@
 #import "MXEnumConstants.h"
 #import "MXRoomMembersCount.h"
 #import "MXStore.h"
+#import "MXRoomTombStoneContent.h"
+#import "MXRoomCreateContent.h"
 
 @class MXSession;
 
@@ -150,6 +152,15 @@ Use MXRoomSummary.displayname to get a computed room display name.
  */
 @property (nonatomic, readonly) NSString *encryptionAlgorithm;
 
+/**
+ Indicate whether this room is obsolete (had a `m.room.tombstone` state event type).
+ */
+@property (nonatomic, readonly) BOOL isObsolete;
+
+/**
+ If any the state event content for event type `m.room.tombstone`
+ */
+@property (nonatomic, strong, readonly) MXRoomTombStoneContent *tombStoneContent;
 
 /**
  Create a `MXRoomState` instance.