+++ MatrixSDK/Data/MXRoomSummary.m
@@ -562,6 +562,8 @@ NSString *const kMXRoomSummaryDidChangeNotification = @"kMXRoomSummaryDidChangeN
         _lastMessageString = lastMessageData[@"lastMessageString"];
         _lastMessageAttributedString = lastMessageData[@"lastMessageAttributedString"];
         _lastMessageOthers = lastMessageData[@"lastMessageOthers"];
+        
+        _hiddenFromUser = [aDecoder decodeBoolForKey:@"hiddenFromUser"];
     }
     return self;
 }
@@ -619,6 +621,8 @@ NSString *const kMXRoomSummaryDidChangeNotification = @"kMXRoomSummaryDidChangeN
     {
         [aCoder encodeObject:lastMessageData forKey:@"lastMessageData"];
     }
+    
+    [aCoder encodeBool:_hiddenFromUser forKey:@"hiddenFromUser"];
 }